PD_3:4/requirements.txt
Django==2.2.*
Pillow==6.2.*
django-import-export==2.0
djangorestframework==3.10.*
gunicorn
django-heroku